requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.1
    # via python-binance
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
asynctest==0.13.0
    # via aiohttp
attrs==22.1.0
    # via aiohttp
backports-zoneinfo==0.2.1 ; python_version < "3.9"
    # via
    #   -r requirements.in
    #   tzlocal
backtesting==0.3.3
    # via -r requirements.in
bokeh==2.4.3
    # via backtesting
certifi==2022.9.14
    # via requests
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via rich-click
colorama==0.4.5
    # via click
commonmark==0.9.1
    # via rich
cycler==0.11.0
    # via matplotlib
dateparser==1.1.1
    # via python-binance
fonttools==4.37.3
    # via matplotlib
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==4.12.0
    # via
    #   click
    #   rich-click
jinja2==3.1.2
    # via bokeh
kiwisolver==1.4.4
    # via matplotlib
markupsafe==2.1.1
    # via jinja2
matplotlib==3.5.3
    # via mplfinance
mplfinance==0.12.9b1
    # via -r requirements.in
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
numpy==1.21.6
    # via
    #   -r requirements.in
    #   backtesting
    #   bokeh
    #   matplotlib
    #   pandas
packaging==21.3
    # via
    #   bokeh
    #   matplotlib
pandas==1.3.5
    # via
    #   -r requirements.in
    #   backtesting
    #   mplfinance
pillow==9.2.0
    # via
    #   bokeh
    #   matplotlib
plotly==5.10.0
    # via -r requirements.in
pygments==2.13.0
    # via rich
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-binance==1.0.16
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   dateparser
    #   matplotlib
    #   pandas
python-dotenv==0.21.0
    # via -r requirements.in
pytz==2022.2.1
    # via
    #   dateparser
    #   pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via bokeh
regex==2022.3.2
    # via dateparser
requests==2.28.1
    # via python-binance
rich==12.5.1
    # via rich-click
rich-click==1.5.2
    # via -r requirements.in
six==1.16.0
    # via
    #   python-binance
    #   python-dateutil
tenacity==8.1.0
    # via plotly
tornado==6.2
    # via bokeh
typing-extensions==4.3.0
    # via
    #   aiohttp
    #   async-timeout
    #   bokeh
    #   importlib-metadata
    #   kiwisolver
    #   rich
    #   yarl
tzdata==2022.2
    # via tzlocal
tzlocal==4.2
    # via dateparser
ujson==5.5.0
    # via python-binance
urllib3==1.26.12
    # via requests
websockets==10.3
    # via python-binance
yarl==1.8.1
    # via aiohttp
zipp==3.8.1
    # via importlib-metadata
